+++ nwAlgorithm.h
@@ -36,6 +36,9 @@ typedef struct Allignment
 /**
  * @brief Makes the matrix to find the best allignment
  * 
+ * The algorithm has computational complexity of O(n^2)
+ * In reality it is O(size of genseq1 * size of genseq2), which can be aproximated do n^2
+ * 
  * @param genseq1 the first genetic sequence
  * @param genseq2 the second genetic sequence
  * @param vector<T_Allignment>& to place the best allignment
@@ -67,4 +70,4 @@ int maxValue(int a, int b, int c);
  * 
  * @param line the string to print
  */
-void printString(string& line, int counter);
+void printString(string& line, long int counter);
